requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
joblib==0.17.0
lyricsgenius==2.0.2
nltk==3.5
psycopg2==2.8.6
pytz==2020.4
regex==2020.11.13
requests==2.25.0
soupsieve==2.0.1
sqlparse==0.4.1
textblob==0.15.3
tqdm==4.54.0
urllib3==1.26.2
whitenoise==5.2.0
